+++ src/align/hts/hts.jl
@@ -7,11 +7,10 @@
 # License is MIT: https://github.com/BioJulia/Bio.jl/blob/master/LICENSE.md
 
 import Bio: Ragel
+import Bio.Intervals: Chunk, BGZFIndex, BinIndex, LinearIndex, PseudoBin, LinearWindowSize
 import BGZFStreams: BGZFStream, VirtualOffset, virtualoffset
 import BufferedStreams   # this is necessary though I don't know why.
 import BufferedStreams: BufferedInputStream
 
-include("chunk.jl")
 include("sam/sam.jl")
 include("bam/bam.jl")
-include("tabix/tabix.jl")
